PIZZA-10 Added authorization
