+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@auth0/auth0-react": "^1.2.0",
     "@hookform/resolvers": "^1.3.2",
     "@reduxjs/toolkit": "^1.5.0",
     "@sentry/react": "^5.29.2",
