+++ src/routers/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import { ReactElement } from 'react';
 import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom';
 
 import RegistrationPage from 'pages/Auth/RegistrationPage';
@@ -7,27 +7,16 @@ import PizzaConstructorPage from 'pages/PizzaConstructorPage';
 import OrderCheckoutPage from 'pages/OrderCheckoutPage';
 import OrderConfirmPage from 'pages/OrderConfirmPage';
 import OrdersListPage from 'pages/OrdersListPage';
-import ProtectedRoute from './ProtectedRoute';
 
-const AppRouter = (): JSX.Element => (
+const AppRouter = (): ReactElement => (
   <BrowserRouter>
     <Switch>
       <Route path="/" component={LogInPage} exact />
-      <ProtectedRoute path="/registration">
-        <RegistrationPage />
-      </ProtectedRoute>
-      <ProtectedRoute path="/home">
-        <PizzaConstructorPage />
-      </ProtectedRoute>
-      <ProtectedRoute path="/order-checkout">
-        <OrderCheckoutPage />
-      </ProtectedRoute>
-      <ProtectedRoute path="/order-confirm/:status?">
-        <OrderConfirmPage />
-      </ProtectedRoute>
-      <ProtectedRoute path="/orders-history">
-        <OrdersListPage />
-      </ProtectedRoute>
+      <Route path="/registration" component={RegistrationPage} />
+      <Route path="/home" component={PizzaConstructorPage} />
+      <Route path="/order-checkout" component={OrderCheckoutPage} />
+      <Route path="/order-confirm/:status?" component={OrderConfirmPage} />
+      <Route path="/orders-history" component={OrdersListPage} />
       <Redirect to="/" />
     </Switch>
   </BrowserRouter>
